move tilde and history back to src/;fix some prototypes

+++ src/tty.c
@@ -2005,7 +2005,7 @@ vcs_io(buf, op)
     char *buf;
     int op;
 {
-    int (*fn)();
+    ssize_t (*fn)();
     int vcsfd, flag;
     char vcs_name[16];
     char vcsa_name[16];
